Firmware_MASTER --> Parametric to any number of servos.
This version of the firmware is parametric such that it will initiate
any given number of servos through a initial synchronization signal from
the GUI. NOTE: This code has errors and will not compile, I'll fix that
soon. NOTE: This code does not account for the subtleties involved in
having >12 servos. With >12 servo's, you will need to attach a servo
shield, which will then involve using I2C communication. To be
continued...